requirements.txt
python-dotenv==1.0.1
rich==13.7.1
tenacity==8.5.0
numpy==1.26.4
pandas==2.2.2
tiktoken==0.7.0
datasets==2.21.0
faiss-cpu==1.8.0.post1

transformers==4.45.0
accelerate==0.33.0

wandb==0.17.5
fastapi==0.111.0
uvicorn[standard]==0.30.3

openai==1.40.2
anthropic==0.39.0
groq==0.9.0

lm-eval==0.4.3
evalplus==0.3.1
